Add pre-commit configuration
Add a pre-commit configuration that checks for old Python syntax and trailing-whitespace. It checks once a month if there are any updates, so it's easy on maintanance.

pre-commit can be enabled on https://pre-commit.ci/.

There are a lot of other pre-commit hooks available that check automatically for things, see https://pre-commit.com/hooks.html.