Update core.py by adding 'objectives_normalized_values' function
When I use platypus.normalize(algorithm.result), it calls the normalize function from playpus.tools and gives error due to 'is_zero' function. To call the actual 'normalize' function, the user needs to use platypus.core.normalize and it returns maximum and minimum interval only with normalized values as attributes of Solution. However, this process is difficult to understand for a user.